Update fix commit 1

+++ Services/Fsel.Identity/Fsel.Identity.Application/Commands/AdminCmd/UpdateStudentByAdminCommand.cs
@@ -109,16 +109,13 @@ namespace Fsel.Identity.Application.Commands.AdminCmd
                     }
                 }
             }
-            else if (student != null)
+            else if (student != null && student.ParentStudents != null && student.ParentStudents.Count > 0)
             {
-                if (student.ParentStudents != null && student.ParentStudents.Count > 0)
+                var human = userView.Human?.Student?.ParentStudents.FirstOrDefault()?.Parent?.Human;
+                if (human != null)
                 {
-                    var human = userView.Human?.Student?.ParentStudents.FirstOrDefault()?.Parent?.Human;
-                    if (human != null)
-                    {
-                        await _humanRepository.DeleteAsync(human);
-                        await _humanRepository.UnitOfWork.SaveEntitiesAsync(cancellationToken).ConfigureAwait(false);
-                    }
+                    await _humanRepository.DeleteAsync(human);
+                    await _humanRepository.UnitOfWork.SaveEntitiesAsync(cancellationToken).ConfigureAwait(false);
                 }
             }
